Renames and disentangles checksum validator classes.
closes #2068
