+++ .rubocop.yml
@@ -79,15 +79,7 @@ RSpec/NamedSubject:
   Enabled: false
 
 RSpec/NestedGroups:
-  Max: 4 # default: 3
-  Exclude:
-    - 'spec/jobs/validate_moab_job_spec.rb'
-    - 'spec/services/audit/catalog_to_moab_spec.rb'
-    - 'spec/services/audit/moab_to_catalog_spec.rb'
-    - 'spec/requests/objects_controller_content_diff_spec.rb'
-    - 'spec/requests/objects_controller_file_spec.rb'
-    - 'spec/services/checksum_validator_spec.rb'
-    - 'spec/services/complete_moab_service/*.rb'
+  Enabled: false
 
 Style/AccessModifierDeclarations:
   Exclude:
