+++ .rubocop_todo.yml
@@ -41,9 +41,9 @@ Metrics/AbcSize:
     - 'app/models/preserved_object.rb'
     - 'app/services/audit/catalog_to_archive.rb'
     - 'app/services/audit/catalog_to_moab.rb'
-    - 'app/services/audit/checksum.rb'
+    - 'app/services/audit/checksum_validator_utils.rb'
     - 'app/services/audit/moab_to_catalog.rb'
-    - 'app/services/checksum_validator.rb'
+    - 'app/services/audit/checksum_validator.rb'
     - 'app/services/complete_moab_service/check_existence.rb'
     - 'app/services/complete_moab_service/update_version.rb'
     - 'app/services/moab_storage_root_report_service.rb'
@@ -94,7 +94,7 @@ RSpec/StubbedMock:
     - 'spec/jobs/moab_to_catalog_job_spec.rb'
     - 'spec/jobs/part_replication_audit_job_spec.rb'
     - 'spec/services/audit/catalog_to_moab_spec.rb'
-    - 'spec/services/checksum_validator_spec.rb'
+    - 'spec/services/audit/checksum_validator_spec.rb'
     - 'spec/services/complete_moab_service/check_existence_spec.rb'
     - 'spec/services/complete_moab_service/create_after_validation_spec.rb'
     - 'spec/services/complete_moab_service/update_version_after_validation_spec.rb'
