+++ README.md
@@ -260,7 +260,7 @@ MoabStorageRoot.find_each { |msr| msr.validate_expired_checksums! }
 #### Single Druid
 Synchronously, from Rails console (will take a long time for very large objects):
 ```ruby
-Audit::Checksum.validate_druid(druid)
+Audit::ChecksumValidatorUtils.validate_druid(druid)
 ```
 
 #### Druid List
@@ -268,7 +268,7 @@ Audit::Checksum.validate_druid(druid)
 
 Synchronously, from Rails console:
 ```ruby
-Audit::Checksum.validate_list_of_druids('/file/path/to/your/csv/druid_list.csv')
+Audit::ChecksumValidatorUtils.validate_list_of_druids('/file/path/to/your/csv/druid_list.csv')
 ```
 
 #### Druids with a particular status on a particular storage root
@@ -276,7 +276,7 @@ Audit::Checksum.validate_list_of_druids('/file/path/to/your/csv/druid_list.csv')
 For example, if you wish to run CV on all the "validity_unknown" druids on storage root 15, from console:
 
 ```ruby
-Audit::Checksum.validate_status_root(:validity_unknown, 'services-disk15')
+Audit::ChecksumValidatorUtils.validate_status_root(:validity_unknown, 'services-disk15')
 ```
 
 [Valid status strings](https://github.com/sul-dlss/preservation_catalog/blob/main/app/models/complete_moab.rb#L1-L10)
